requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.4
django-ckeditor==6.7.1
django-js-asset==2.2.0
gunicorn==21.2.0
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
